move and refactor common RPC code
