+++ src/rpc/common/command_decorators.cpp
@@ -1,0 +1,11 @@
+#include "command_decorators.h"
+
+namespace cryptonote::rpc {
+
+void RPC_COMMAND::set_bt() {
+  bt = true;
+  response_b64.format = json_binary_proxy::fmt::bt;
+  response_hex.format = json_binary_proxy::fmt::bt;
+}
+
+} // namespace cryptonote::rpc
