+++ src/rpc/common/rpc_binary.cpp
