+++ src/rpc/common/rpc_binary.h
